+++ src/main/java/com/gabrielosorio/gestor_inteligente/model/Supplier.java
@@ -3,19 +3,19 @@ package com.gabrielosorio.gestor_inteligente.model;
 import com.gabrielosorio.gestor_inteligente.model.enums.PaymentMethod;
 
 public class Supplier {
-    private Integer id;
+    private long id;
     private String name;
 
-    public Supplier(Integer id, String name) {
+    public Supplier(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
